+++ src/main/java/org/springframework/samples/petclinic/rest/deserializer/JacksonCustomOwnerDeserializer.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.springframework.samples.petclinic.rest;
+package org.springframework.samples.petclinic.rest.deserializer;
 
 import java.io.IOException;
