+++ src/main/java/org/springframework/samples/petclinic/rest/deserializer/JacksonCustomVisitDeserializer.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.springframework.samples.petclinic.rest;
+package org.springframework.samples.petclinic.rest.deserializer;
 
 import java.io.IOException;
 import java.text.ParseException;
