+++ src/main/java/org/springframework/samples/petclinic/rest/serializer/JacksonCustomVisitSerializer.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.springframework.samples.petclinic.rest;
+package org.springframework.samples.petclinic.rest.serializer;
 
 import java.io.IOException;
 import java.text.Format;
